refactor(database): initial changes for RESTful standard
