+++ database/src/business/index.js
@@ -3,3 +3,4 @@ export { default as NftService } from './nft.service';
 export { default as NftCommitmentService } from './nft-commitment.service';
 export { default as FtService } from './ft.service';
 export { default as FtCommitmentService } from './ft-commitment.service';
+export { default as UserService } from './user.service';
