+++ database/src/routes/index.js
@@ -3,3 +3,4 @@ export { default as initializeNftRoutes } from './nft.routes';
 export { default as initializeNftCommitmentRoutes } from './nft-commitment.routes';
 export { default as initializeFtRoutes } from './ft.routes';
 export { default as initializeFtCommitmentRoutes } from './ft-commitment.routes';
+export { default as initializeUserRoutes } from './user.routes';
